pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "audiomason"
version = "0.1.0"
description = "Deterministic ASCII-only CLI for importing/normalizing/tagging audiobooks"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0.1",
  "mutagen>=1.47.0",
]

[project.scripts]
audiomason = "audiomason.cli:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["audiomason*"]
